@@ -1,55 +1,46 @@
 {
-  "name": "typescript-and-chill",
-  "version": "2.1.2",
-  "description": "A CLI tool to automate setting up a TypeScript developer environment",
-  "main": "dist/index.js",
-  "bin": {
-    "typescript-and-chill": "dist/index.js"
-  },
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/",
-    "templates/"
-  ],
-  "scripts": {
-    "test": "jest",
-    "lint": "eslint 'src/**/*.ts'",
-    "format": "prettier --write 'src/**/*.ts'",
-    "release": "node scripts/update-changelog.js",
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "ts-node src/index.ts",
-    "prepare": "npm run build"
-  },
-  "dependencies": {
-    "child_process": "^1.0.2",
-    "esbuild": "^0.15.0",
-    "eslint": "^9.19.0",
-    "jest": "^29.0.0",
-    "prettier": "^3.4.2"
-  },
-  "devDependencies": {
-    "@types/node": "^22.12.0",
-    "@typescript-eslint/eslint-plugin": "^8.22.0",
-    "@typescript-eslint/parser": "^8.22.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-prettier": "^5.2.3",
-    "ts-jest": "^29.0.0",
-    "typescript": "^5.7.3"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/theleepriest/typescript-and-chill.git"
-  },
-  "keywords": [
-    "cli",
-    "setup",
-    "typescript",
-    "eslint",
-    "prettier",
-    "aws",
-    "cdk"
-  ],
-  "author": "Lee Priest",
-  "license": "MIT"
+	"name": "typescript-and-chill",
+	"version": "2.1.2",
+	"description": "A CLI tool to automate setting up a TypeScript developer environment",
+	"main": "dist/index.js",
+	"bin": {
+		"typescript-and-chill": "dist/index.js"
+	},
+	"types": "dist/index.d.ts",
+	"files": ["dist/", "templates/"],
+	"scripts": {
+		"test": "vitest",
+		"lint": "eslint 'src/**/*.ts'",
+		"format": "prettier --write 'src/**/*.ts'",
+		"release": "ts-node scripts/update-changelog.js",
+		"build": "tsc",
+		"start": "ts-node dist/index.js",
+		"dev": "ts-node src/index.ts",
+		"prepare": "npm run build"
+	},
+	"dependencies": {
+		"child_process": "^1.0.2",
+		"esbuild": "^0.25.4"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "1.9.4",
+		"@types/node": "^22.15.18",
+		"typescript": "^5.8.3",
+		"vitest": "^3.1.3"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/theleepriest/typescript-and-chill.git"
+	},
+	"keywords": [
+		"cli",
+		"setup",
+		"typescript",
+		"eslint",
+		"prettier",
+		"aws",
+		"cdk"
+	],
+	"author": "Lee Priest",
+	"license": "MIT"
 }
