@@ -1,7 +1,7 @@
-import fs from 'fs';
-import path from 'path';
+import fs from "node:fs";
+import path from "node:path";
 
 export const setupGitIgnore = (): void => {
-  console.log('🛠️ Setting up Gitignore...');
-  fs.copyFileSync(path.join(__dirname, '../templates/gitignore'), '.gitignore');
+	console.log("🛠️ Setting up Gitignore...");
+	fs.copyFileSync(path.join(__dirname, "../templates/gitignore"), ".gitignore");
 };
