@@ -1,5 +1,5 @@
 export const toPascalCase = (stringToFix: string): string => {
-  return stringToFix.replace(/(^\w|[-_]\w)/g, (match) =>
-    match.replace(/[-_]/, '').toUpperCase()
-  );
+	return stringToFix.replace(/(^\w|[-_]\w)/g, (match) =>
+		match.replace(/[-_]/, "").toUpperCase(),
+	);
 };
